+++ beem/__init__.py
@@ -4,7 +4,6 @@ from .hive import Hive
 from .version import version as __version__
 __all__ = [
     "steem",
-    "aes",
     "account",
     "amount",
     "asset",