+++ beemgraphenebase/__init__.py
@@ -8,6 +8,7 @@ from .version import version as __version__
 # from . import dictionary as BrainKeyDictionary
 
 __all__ = ['account',
+           'aes',
            'base58',
            'bip32'
            'bip38',