+++ tests/beem/test_wallet.py
@@ -138,30 +138,3 @@ class Testcases(unittest.TestCase):
         ):
             self.wallet.getPostingKeysForAccount("test")
 
-    def test_encrypt(self):
-        stm = self.stm
-        self.wallet.steem = stm
-        self.wallet.unlock(pwd="TestingOneTwoThree")
-        self.wallet.masterpassword = "TestingOneTwoThree"
-        self.assertEqual([self.wallet.encrypt_wif("5HqUkGuo62BfcJU5vNhTXKJRXuUi9QSE6jp8C3uBJ2BVHtB8WSd"),
-                          self.wallet.encrypt_wif("5KN7MzqK5wt2TP1fQCYyHBtDrXdJuXbUzm4A9rKAteGu3Qi5CVR")],
-                         ["6PRN5mjUTtud6fUXbJXezfn6oABoSr6GSLjMbrGXRZxSUcxThxsUW8epQi",
-                          "6PRVWUbkzzsbcVac2qwfssoUJAN1Xhrg6bNk8J7Nzm5H7kxEbn2Nh2ZoGg"])
-        self.wallet.masterpassword = "Satoshi"
-        self.assertEqual([self.wallet.encrypt_wif("5HtasZ6ofTHP6HCwTqTkLDuLQisYPah7aUnSKfC7h4hMUVw2gi5")],
-                         ["6PRNFFkZc2NZ6dJqFfhRoFNMR9Lnyj7dYGrzdgXXVMXcxoKTePPX1dWByq"])
-        self.wallet.masterpassword = "TestingOneTwoThree"
-
-    def test_deencrypt(self):
-        stm = self.stm
-        self.wallet.steem = stm
-        self.wallet.unlock(pwd="TestingOneTwoThree")
-        self.wallet.masterpassword = "TestingOneTwoThree"
-        self.assertEqual([self.wallet.decrypt_wif("6PRN5mjUTtud6fUXbJXezfn6oABoSr6GSLjMbrGXRZxSUcxThxsUW8epQi"),
-                          self.wallet.decrypt_wif("6PRVWUbkzzsbcVac2qwfssoUJAN1Xhrg6bNk8J7Nzm5H7kxEbn2Nh2ZoGg")],
-                         ["5HqUkGuo62BfcJU5vNhTXKJRXuUi9QSE6jp8C3uBJ2BVHtB8WSd",
-                          "5KN7MzqK5wt2TP1fQCYyHBtDrXdJuXbUzm4A9rKAteGu3Qi5CVR"])
-        self.wallet.masterpassword = "Satoshi"
-        self.assertEqual([self.wallet.decrypt_wif("6PRNFFkZc2NZ6dJqFfhRoFNMR9Lnyj7dYGrzdgXXVMXcxoKTePPX1dWByq")],
-                         ["5HtasZ6ofTHP6HCwTqTkLDuLQisYPah7aUnSKfC7h4hMUVw2gi5"])
-        self.wallet.masterpassword = "TestingOneTwoThree"